🏗️ イベントハンドラーをTypeScriptで実装し、guildCreate、guildDelete、interactionCreate、readyイベントを追加
